Add constraint pod events in Archiver
Summary:
For long running services, pod events will continue to aggregate and in a scenario of crash loop, it can lead to unnecessary load on Cassandra on read.

To constraint the number of runs for each instances in pod events table, this diff adds a constraint to persist maximum 100 most recent runs in DB.

To constraint #runs in DB, it uses tasksvc to get most recent pod event run_id and then calls tasksvc's delete pod events for most recent pod run - 100.

Test Plan:
pcluster
preprod

Reviewers: mu, adityacb, apoorvaj, #peloton, O521 Project Peloton: Add reviewers

Reviewed By: adityacb, #peloton, O521 Project Peloton: Add reviewers

Subscribers: jenkins

Maniphest Tasks: T1838557

Differential Revision: https://code.uberinternal.com/D2035701

+++ jobmgr/cached/task.go
@@ -122,7 +122,7 @@ func validateMesosTaskID(mesosTaskID, prevMesosTaskID string) bool {
 		return true
 	}
 
-	var newRunID, currentRunID uint32
+	var newRunID, currentRunID uint64
 	var err error
 
 	if newRunID, err = util.ParseRunID(mesosTaskID); err != nil {
